Separate unnecessary dependencies that get pulled in by Wayland testing.
