+++ tests/acceptance-tests/test_nested_mir.cpp
@@ -45,6 +45,7 @@
 #include "mir_test_framework/any_surface.h"
 #include "mir_test_framework/fake_input_device.h"
 #include "mir_test_framework/observant_shell.h"
+#include "mir_test_framework/passthrough_tracker.h"
 #include "mir/test/signal.h"
 #include "mir/test/spin_wait.h"
 #include "mir/test/display_config_matchers.h"
