+++ tests/mir_test_framework/headless_display_buffer_compositor_factory.cpp
@@ -17,7 +17,7 @@
  */
 
 #include "mir_test_framework/headless_display_buffer_compositor_factory.h"
-#include "mir_test_framework/headless_nested_server_runner.h"
+#include "mir_test_framework/passthrough_tracker.h"
 #include "mir/renderer/gl/render_target.h"
 #include "mir/renderer/gl/texture_source.h"
 #include "mir/graphics/display_buffer.h"
